Restructures the file names to better match the overall file structure
